refactor: move bcrypt-adapter to a folder
